requirements.txt
mcp[cli]
uvicorn
python-dotenv